Had to change the way api.js and store works...

+++ src/Helpers/api.js
@@ -1,5 +1,7 @@
 
-import { axiosApi } from '../Api/init'
+import { axiosApi, setJwt } from '../Api/init'
+import decodeJWT from 'jwt-decode'
+import store from '../Redux/store'
 
 
 // This class is a helper method that is useable by all the different components in sending api requests to the backend. It is adaptable to each model endpoint by creating a new instance of the class object and initialising it with the endpoint that relates the endpoint you want to call on. It operates on all CRUD options. 
@@ -13,35 +15,59 @@ class ApiMethods {
 
 // CRUD - CREATE
     create = async (obj, func) => {
-        console.log('In Real Create')
         // "obj" is the new object you are creating
         func = func || null
+        const response = await axiosApi.post(this.endpoint, obj)
 
-        return axiosApi.post(
-            this.endpoint, obj
-        )
+         store.dispatch({
+            type: this.endpoint,
+            customAction: 'create',
+            newlyCreated: response.data
+        })
+
+        func(response.data)
     }
 
 // CRUD - READ-ALL
-    all = async () => {
-        const array = await axiosApi.get(this.endpoint)
-        return array
+    all = async (func) => {
+        const response = await axiosApi.get(this.endpoint)
+
+        store.dispatch({
+            type: this.endpoint,
+            customAction: 'get_all',
+            all: response.data
+        })
+
+        func(response.data)
     }
 
  // CRUD - READ-SINGLE (findById)
-    find = async (id) => {
-        const object = await axiosApi.get(this.endpoint + id)
-        return object
-    }
+    // find = async (id) => {
+    //     const object = await axiosApi.get(this.endpoint + `/${id}`)
+    //     return object
+    // }
 
  // CRUD - UPDATE
-    update = (id, obj, func) => {
+    update = async (id, obj, func) => {
         // The middle paramater here passed as "obj" are the fields that are being updated.
-        func = func || null
-        
-        axiosApi.put(
-            this.endpoint + id, obj
-        ).then(func)
+        try {
+            func = func || null
+            
+            const response = await axiosApi.put(this.endpoint + `/${id}`, obj)
+            console.log(response)
+    
+            store.dispatch({
+                type: this.endpoint,
+                customAction: 'update',
+                updated: response.data, 
+                id: id
+            })
+    
+            func()
+
+        } catch (error) {
+            console.log(error.response)
+        }
     }
 
 // CRUD - DESTROY
@@ -49,12 +75,47 @@ class ApiMethods {
         // The middle paramater here passes the objects "ownership" id... Consider it as a foreign key e.g. user_id
         func = func || null
 
-        axiosApi.delete(
-            this.endpoint + id, {
-                data: { id: authorized.userId }
-            }
-        ).then(func)       
+        axiosApi.delete(this.endpoint + `/${id}`, {data: { id: authorized.userId }})  
+        
+        store.dispatch({
+            type: this.endpoint,
+            customAction: 'delete',
+            id: id
+        })
+
+        func()
     }
+
+    authenticate = async (newUser, url, func) => {
+        try {
+
+
+            const response = await axiosApi.post(`/users${url}`, newUser)
+
+            const token = response.data.token
+            const user = token && decodeJWT(token)
+            user.token = token
+
+            setJwt(user)
+
+            store.dispatch({
+                type: 'set_loggedIn',
+                loggedIn: true,
+                currentUser: user
+            })
+
+            func(user)
+
+          
+        } catch (error) {
+            store.dispatch({
+                type: 'set_loginError',
+                loginError: error.response.data
+            })
+        }
+
+    }
+    
 }
 
 
@@ -65,4 +126,4 @@ const user = new ApiMethods('/users')
 // const products = new ApiMethods('http://localhost:3001/delivery-day')
 
 
-export default { orders, products, company }
+export default { orders, products, company, user }
